ldap: filter out group-managed service accounts
For some reason, when doing an LDAP search for (objectClass=computer),
we get both computers and group managed service accounts.

This is mostly harmless, as the service accounts aren't present in
DNS so Teleport just ignores them and only registers the actual
computer objects. It does, however, log a DNS error message that
could be interpreted as an actual problem, so to be safe we explicitly
filter these results out from the search.

This was discovered when testing with AWS managed Active Directory,
which creates a gMSA for DNS.